chore: add readme

+++ conda_poetry_liaison/clean.py
@@ -5,6 +5,8 @@ from .utils import get_files
 
 
 def run_cleaning(env_path):
+    if not isinstance(env_path, str):
+        raise TypeError(f"env_path should be a string, not {type(env_path)}")
     direct_url_paths = get_files(env_path, "direct_url.json")
     for path in direct_url_paths:
         os.remove(path)
